saving progress on the hotkey refactor
